Move shared/IPC.ts to main.
